requirements.txt
Django>=4.2.0
supabase>=2.0.0
Pillow>=10.0.0
django-cors-headers>=4.0.0
python-decouple>=3.8
django-storages>=1.14.0
boto3>=1.34.0
psycopg2-binary>=2.9.0
django-ckeditor>=6.7.0
django-taggit>=4.0.0